+++ Jmol/src/org/openscience/jmol/LineShape.java
@@ -37,8 +37,8 @@ class LineShape implements Shape, Transformable {
 
   }
 
-  public double getZ() {
-    return screenPositionEnd.z + 4.0;
+  public int getZ() {
+    return (int)screenPositionEnd.z + 4;
   }
 
   /**
